rft(doc-analyzer-backend): use float for judges score

+++ services/doc-analyzer-backend/api/models/analisys/result_data.py
@@ -5,6 +5,6 @@ from api.models.analisys.result_iter_data import ResultIterData
 
 class ResultData(BaseModel):
     answer: str
-    score: int | None = None
+    score: float | None = None
     judgement: str | None = None
     answer_iterations: list[ResultIterData] = None
